@@ -1,9 +1,9 @@
 //go:build go1.22
 
-// Package openapi provides primitives to interact with the openapi HTTP API.
+// Package gen provides primitives to interact with the openapi HTTP API.
 //
 // Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
-package openapi
+package gen
 
 import (
 	"bytes"
