@@ -95,14 +95,14 @@ func (sc StreamChatImpl) Execute(ctx context.Context, userMessage string, onEven
 	}
 
 	// Load prior conversation to preserve context
-	history, _, err := sc.chatMessageRepo.ListChatMessages(spanCtx, 0) // full history (or paginated by repo)
+	history, _, err := sc.chatMessageRepo.ListChatMessages(spanCtx, 15)
 	if tracing.RecordErrorAndStatus(span, err) {
 		return err
 	}
 
 	// Build chat request: system + history (excluding old system messages) + current user turn
 	messages := make([]domain.LLMChatMessage, 0, len(systemPrompt)+len(history)+1)
-	messages = append(messages, systemPrompt...) // Fresh system prompt
+	messages = append(messages, systemPrompt...)
 
 	for _, msg := range history {
 		// Skip old system messages to avoid stale todo data
@@ -188,7 +188,7 @@ func (sc StreamChatImpl) Execute(ctx context.Context, userMessage string, onEven
 	userMsg := domain.ChatMessage{
 		ID:             userMessageID,
 		ConversationID: domain.GlobalConversationID,
-		ChatRole:       domain.ChatRole("user"),
+		ChatRole:       domain.ChatRole_User,
 		Content:        userMessage,
 		Model:          req.Model,
 		CreatedAt:      time.Now().UTC(),
@@ -202,7 +202,7 @@ func (sc StreamChatImpl) Execute(ctx context.Context, userMessage string, onEven
 	assistantMsg := domain.ChatMessage{
 		ID:             assistantMessageID,
 		ConversationID: domain.GlobalConversationID,
-		ChatRole:       domain.ChatRole("assistant"),
+		ChatRole:       domain.ChatRole_Assistant,
 		Content:        fullContent.String(),
 		Model:          req.Model,
 		CreatedAt:      time.Now().UTC(),
